+++ domain/src/main/java/com/albertmiro/domain/repository/TripsRepository.kt
@@ -5,5 +5,5 @@ import io.reactivex.Single
 
 interface TripsRepository {
     fun getTrips(foreRefresh: Boolean = false): Single<List<Trip>>
-    fun getTripDetail(id: Int): Single<Trip>
+    fun getTripDetail(id: Long): Single<Trip>
 }
